Move tool nav to separate component
