+++ src/components/tools/Partials/index.js
@@ -4,3 +4,4 @@ export { default as Header } from "./Header.svelte";
 export { default as Help } from "./Help.svelte";
 export { default as ChangeLocation } from "./ChangeLocation.svelte";
 export { default as DownloadChart } from "./DownloadChart.svelte";
+export { default as ToolNavigation } from "./ToolNavigation.svelte";
